+++ WEB_BE/flaskapp/route.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+##
+#  @package flaskapp.route
+#  @file __init__.py
+#  @author  Sinduy
+#  This module contains the route for the Flask app.
+
+from flask import render_template
+from flaskapp import app
+
+@app.route('/')
+def index():
+    return render_template('index.html')
