+++ WEB_BE/run.py
@@ -1,13 +1,7 @@
-from flask import Flask, render_template
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
 
-app = Flask(__name__,
-            static_url_path='', 
-            static_folder='../WEB_FE',
-            template_folder='../WEB_FE')
-
-@app.route('/')
-def index():
-    return render_template('index.html')
+from flaskapp import app
 
 if __name__ == '__main__':
     app.run()
